fix: Unify tradingPrice and marketTotalAmount type to BigDecimal
- 팀 협의 결과 UUID 기반 설계 방향으로 확정
- IndexData update() 메서드 tradingPrice, marketTotalAmount 파라미터
  타입 Long -> BigDecimal로 변경
- new BigDecimal() 변환 제거, BigDecimal 직접 할당으로 수정

+++ src/main/java/com/sprint/mission/findex/domain/indexdata/dto/IndexDataUpdateRequest.java
@@ -25,8 +25,8 @@ public record IndexDataUpdateRequest(
     Long tradingQuantity,
 
     @PositiveOrZero
-    Long tradingPrice,
+    BigDecimal tradingPrice,
 
     @PositiveOrZero
-    Long marketTotalAmount
+    BigDecimal marketTotalAmount
 ) {}
